Add-on now inserts keyframes on leftmouse press.
As a substitute until someone comes along and fixes the way VSE Transform Tool add-on works with automatic keyframing, I made a fork of the add-on that inserts keyframes on all instances of leftmouse press (and leftmouse release and numpad enter) within the add-on. Essentially, the changes present in this fork makes it so whenever a user commits a change they wish to keep, the add-on will insert a keyframe for a strip's location, rotation, scale and crop values on the current frame.

Also, I don't think it needs to be said, but please don't actually copy this into the master branch as this is very much a temporary fix I butchered together.